backend/requirements_render.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
python-dotenv==1.2.1
bcrypt==4.1.3
PyJWT==2.10.1
email-validator==2.3.0
python-multipart==0.0.20
aiohttp==3.13.2
PyPDF2==3.0.1
dnspython==2.8.0
